+++ pkgs/test/lib/src/executable.dart
@@ -9,6 +9,7 @@ import 'package:test_core/src/runner/hack_register_platform.dart'; // ignore: im
 
 import 'runner/browser/platform.dart';
 import 'runner/node/platform.dart';
+import 'runner/wasm/platform.dart';
 
 void main(List<String> args) async {
   registerPlatformPlugin([Runtime.nodeJS], () => NodePlatform());
@@ -18,6 +19,9 @@ void main(List<String> args) async {
     Runtime.safari,
     Runtime.internetExplorer
   ], () => BrowserPlatform.start());
+  registerPlatformPlugin([
+    Runtime.experimentalChromeWasm,
+  ], () => BrowserWasmPlatform.start());
 
   await executable.main(args);
 }